+++ LinkEngine/Entities/Player.cs
@@ -1,20 +1,38 @@
-﻿using LinkEngine.Rendering;
+﻿using LinkEngine.Gameplay.Items;
+using LinkEngine.Gameplay.Quests;
+using LinkEngine.WorldGen;
 using System.Collections.Generic;
-using System.Drawing;
 
 namespace LinkEngine.Entities
 {
     public class Player : Entity
     {
+        System.Random rand = new System.Random();
+
         /// <summary>
         /// The inventory of the player can be accessed by any gametype player class that inherits this class.
         /// </summary>
         public List<InventoryItem> Inventory;
+        public Equipment[] Equipment { get; set; }
+        public List<PlayerQuest> Quests { get; set; }
 
         /// <summary>
-        /// MapLoc is the position of the map according to the camera
+        /// The current level of the Player
         /// </summary>
-        public Point MapLoc { get; set; }
+        public int Level { get; set; }
+        /// <summary>
+        /// The amassed experience of the Player
+        /// </summary>
+        public int Exp { get; set; }
+        /// <summary>
+        /// the required exp to level up
+        /// </summary>
+        public int MaxExp { get; set; }
+
+        /// <summary>
+        /// The location the Player is currently in
+        /// </summary>
+        public Location CurrentLocation { get; set; }
 
         /// <summary>
         /// The camera object is what will be used to hold the image the player can see
@@ -41,18 +59,9 @@ namespace LinkEngine.Entities
         /// <param name="y">the amount to shift the map on the Y axis</param>
         public void ShiftMap(int x, int y)
         {
-            MapLoc = new Point(MapLoc.X + x, MapLoc.Y + y);
+            // MapLoc = new Point(MapLoc.X + x, MapLoc.Y + y);
         }
-        /// <summary>
-        /// SetMap returns the drawn map to use in pictureboxes
-        /// </summary>
-        /// <param name="Width"></param>
-        /// <param name="Height"></param>
-        /// <returns></returns>
-        public void SetMap(int Width, int Height)
-        {
-            camera.View = ScreenObject.Draw(Width, Height, new Point(MapLoc.X * 32, (MapLoc.Y - 10) * 32), null);
-        }
+
         /// <summary>
         /// CalibrateMap will adjust the map location so that the player can be centered on the screen
         /// </summary>
@@ -60,13 +69,26 @@ namespace LinkEngine.Entities
         /// <param name="centerY"></param>
         void CalibrateMap(int centerX, int centerY)
         {
-            MapLoc = new Point(
-                    // X Coordinate
-                    centerX - camera.X,
+            //MapLoc = new Point(
+            //        // X Coordinate
+            //        centerX - camera.X,
 
-                    // Y Coordinate
-                    centerY - camera.Y
-                );
+            //        // Y Coordinate
+            //        centerY - camera.Y
+            //    );
+        }
+
+        public void giveExperience(int exp)
+        {
+            Exp += exp;
+        }
+        public void LevelUp()
+        {
+            if (Exp >= MaxExp)
+            {
+                Level++;
+                MaxExp += rand.Next(100, MaxExp);
+            }
         }
 
         /// <summary>
@@ -155,6 +177,304 @@ namespace LinkEngine.Entities
                     return; // We added the item, and are done, so get out of this function
                 }
             }
+        }
+
+        public bool HasThisQuest(Quest quest)
+        {
+            foreach (PlayerQuest playerQuest in Quests)
+            {
+                if (playerQuest.Details.ID == quest.ID)
+                {
+                    return true;
+                }
+            }
+
+            return false;
+        }
+        public bool CompletedThisQuest(Quest quest)
+        {
+            foreach (PlayerQuest playerQuest in Quests)
+            {
+                if (playerQuest.Details.ID == quest.ID)
+                {
+                    return playerQuest.IsCompleted;
+                }
+            }
+
+            return false;
+        }
+        public bool HasAllQuestCompletionItems(Quest quest)
+        {
+            // See if the player has all the items needed to complete the quest here
+            foreach (QuestCompletionItem qci in quest.QuestCompletionItems)
+            {
+                bool foundItemInPlayersInventory = false;
+
+                // Check each item in the player's inventory, to see if they have it, and enough of it
+                foreach (InventoryItem ii in Inventory)
+                {
+                    if (ii.Details.ID == qci.Details.ID) // The player has the item in their inventory
+                    {
+                        foundItemInPlayersInventory = true;
+
+                        if (ii.Quantity < qci.Quantity) // The player does not have enough of this item to complete the quest
+                        {
+                            return false;
+                        }
+                    }
+                }
+
+                // The player does not have any of this quest completion item in their inventory
+                if (!foundItemInPlayersInventory)
+                {
+                    return false;
+                }
+            }
+
+            // If we got here, then the player must have all the required items, and enough of them, to complete the quest.
+            return true;
+        }
+        public void RemoveQuestCompletionItems(Quest quest)
+        {
+            foreach (QuestCompletionItem qci in quest.QuestCompletionItems)
+            {
+                foreach (InventoryItem ii in Inventory)
+                {
+                    if (ii.Details.ID == qci.Details.ID)
+                    {
+                        // Subtract the quantity from the player's inventory that was needed to complete the quest
+                        ii.Quantity -= qci.Quantity;
+                        break;
+                    }
+                }
+            }
+        }
+        public void MarkQuestCompleted(Quest quest)
+        {
+            // Find the quest in the player's quest list
+            foreach (PlayerQuest pq in Quests)
+            {
+                if (pq.Details.ID == quest.ID)
+                {
+                    // Mark it as completed
+                    pq.IsCompleted = true;
+
+                    return; // We found the quest, and marked it complete, so get out of this function
+                }
+            }
+        }
+        public void RecieveQuest(NPC npc)
+        {
+            // See if the player already has the quest, and if they've completed it
+            bool playerAlreadyHasQuest = false;
+            bool playerAlreadyCompletedQuest = false;
+
+            foreach (PlayerQuest playerQuest in Quests)
+            {
+                if (playerQuest.Details.ID == npc.QuestAvailableHere.ID)
+                {
+                    playerAlreadyHasQuest = true;
+
+                    if (playerQuest.IsCompleted)
+                    {
+                        playerAlreadyCompletedQuest = true;
+                    }
+                }
+            }
+
+            // See if the player already has the quest
+            if (playerAlreadyHasQuest)
+            {
+                // If the player has not completed the quest yet
+                if (!playerAlreadyCompletedQuest)
+                {
+                    // See if the player has all the items needed to complete the quest
+                    bool playerHasAllItemsToCompleteQuest = true;
+
+                    foreach (QuestCompletionItem qci in npc.QuestAvailableHere.QuestCompletionItems)
+                    {
+                        bool foundItemInPlayersInventory = false;
+
+                        // Check each item in the player's inventory, to see if they have it, and enough of it
+                        foreach (InventoryItem ii in Inventory)
+                        {
+                            // The player has this item in their inventory
+                            if (ii.Details.ID == qci.Details.ID)
+                            {
+                                foundItemInPlayersInventory = true;
+
+                                if (ii.Quantity < qci.Quantity)
+                                {
+                                    // The player does not have enough of this item to complete the quest
+                                    playerHasAllItemsToCompleteQuest = false;
+
+                                    // There is no reason to continue checking for the other quest completion items
+                                    break;
+                                }
+
+                                // We found the item, so don't check the rest of the player's inventory
+                                break;
+                            }
+                        }
+
+                        // If we didn't find the required item, set our variable and stop looking for other items
+                        if (!foundItemInPlayersInventory)
+                        {
+                            // The player does not have this item in their inventory
+                            playerHasAllItemsToCompleteQuest = false;
+
+                            // There is no reason to continue checking for the other quest completion items
+                            break;
+                        }
+                    }
+
+                    // The player has all items required to complete the quest
+                    if (playerHasAllItemsToCompleteQuest)
+                    {
+                        // Display message
+                        // World.Output.Text += Environment.NewLine;
+                        // World.Output.Text += "You complete the '" + npc.QuestAvailableHere.Name + "' quest." + Environment.NewLine;
+
+                        // Remove quest items from inventory
+                        foreach (QuestCompletionItem qci in npc.QuestAvailableHere.QuestCompletionItems)
+                        {
+                            foreach (InventoryItem ii in Inventory)
+                            {
+                                if (ii.Details.ID == qci.Details.ID)
+                                {
+                                    // Subtract the quantity from the player's inventory that was needed to complete the quest
+                                    ii.Quantity -= qci.Quantity;
+                                    break;
+                                }
+                            }
+                        }
+
+                        // Give quest rewards
+                        giveExperience(npc.QuestAvailableHere.RewardExperiencePoints);
+
+                        // Add the reward item to the player's inventory
+                        bool addedItemToPlayerInventory = false;
+
+                        foreach (InventoryItem ii in Inventory)
+                        {
+                            if (ii.Details.ID == npc.QuestAvailableHere.RewardItem.ID)
+                            {
+                                // They have the item in their inventory, so increase the quantity by one
+                                ii.Quantity++;
+
+                                addedItemToPlayerInventory = true;
+
+                                break;
+                            }
+                        }
+
+                        // They didn't have the item, so add it to their inventory, with a quantity of 1
+                        if (!addedItemToPlayerInventory)
+                        {
+                            Inventory.Add(new InventoryItem(npc.QuestAvailableHere.RewardItem, 1));
+                        }
+
+                        // Mark the quest as completed
+                        // Find the quest in the player's quest list
+                        foreach (PlayerQuest pq in Quests)
+                        {
+                            if (pq.Details.ID == npc.QuestAvailableHere.ID)
+                            {
+                                // Mark it as completed
+                                pq.IsCompleted = true;
+
+                                break;
+                            }
+                        }
+                    }
+                }
+            }
+            else
+            {
+                // The Player does not already have the quest
+                // Add the quest to the player's quest list
+                Quests.Add(new PlayerQuest(npc.QuestAvailableHere));
+            }
+            //World.HUD.Update();
+        }
+
+
+        public bool HasRequiredItemToEnterThisLocation(Location location)
+        {
+            if (location.ItemRequiredToEnter == null)
+            {
+                // There is no required item for this location, so return "true"
+                return true;
+            }
+
+            // See if the player has the required item in their inventory
+            foreach (InventoryItem ii in Inventory)
+            {
+                if (ii.Details.ID == location.ItemRequiredToEnter.ID)
+                {
+                    // We found the required item, so return "true"
+                    return true;
+                }
+            }
+
+            // We didn't find the required item in their inventory, so return "false"
+            return false;
+        }
+        public void MoveTo(Location newLocation)
+        {
+            //Does the location have any required items
+            if (newLocation.ItemRequiredToEnter != null)
+            {
+                // See if the player has the required item in their inventory
+                bool playerHasRequiredItem = false;
+
+                foreach (InventoryItem ii in Inventory)
+                {
+                    if (ii.Details.ID == newLocation.ItemRequiredToEnter.ID)
+                    {
+                        // We found the required item
+                        playerHasRequiredItem = true;
+                        break; // Exit out of the foreach loop
+                    }
+                }
+
+                if (!playerHasRequiredItem)
+                {
+                    // We didn't find the required item in their inventory, so display a message and stop trying to move
+                    return;
+                }
+            }
+            CurrentLocation = newLocation;
+            try
+            {
+                CurrentLocation.MonsterLivingHere = newLocation.MonsterLivingHere;
+            }
+            catch { CurrentLocation.MonsterLivingHere = null; }
+
+            //World.HUD.UpdateNPCs();
+        }
+
+        public InventoryItem ItemByName(string name)
+        {
+            foreach (InventoryItem ii in Inventory)
+            {
+                if (ii.Details.Name == name)
+                {
+                    return ii;
+                }
+            }
+            return null;
+        }
+        public Equipment EquipmentByName(string name)
+        {
+            foreach (Equipment equ in Equipment)
+            {
+                if (equ.Name == name)
+                {
+                    return equ;
+                }
+            }
+            return null;
         }
     }
 }
