+++ LinkEngine/Gameplay/Combat/Combat.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿using LinkEngine.Entities;
+using LinkEngine.Gameplay.Modifiers;
+using System;
 using System.Collections.Generic;
 
 namespace LinkEngine.RPG
@@ -45,19 +47,19 @@ namespace LinkEngine.RPG
         /// </summary>
         /// <param name="Attacker">The entity executing the command</param>
         /// <param name="Defender">The target entity</param>
-        public void MonsterAttack (Monster Attacker, RPGCharacter Defender)
+        public void MonsterAttack (Enemy Attacker, Player Defender)
         {
             // Check if the defender has an agility or luck Ability
             if (Defender.Luck > 0 || Defender.Agility > 0)
             {
                 // check for any special abilities
                 // if the luck or agility Ability is high enough, dodge the attack
-                canDodge = Defender.LuckCheck(Attacker.Strength, Defender.LuckModifiers.ToArray());
-                canDodge = Defender.AgilityCheck(Attacker.Strength, Defender.AgilityModifiers.ToArray());
+                canDodge = Defender.RPG.LuckCheck(Attacker.Strength, Defender.RPG.LuckModifiers.ToArray());
+                canDodge = Defender.RPG.AgilityCheck(Attacker.Strength, Defender.RPG.AgilityModifiers.ToArray());
 
                 // if the defender has a higher strength or endurance, he can block the attack
-                canBlock = Defender.StrengthCheck(Attacker.Strength, Defender.StrengthModifiers.ToArray());
-                canBlock = Defender.EnduranceCheck(Attacker.Strength, Defender.EnduranceModifiers.ToArray());
+                canBlock = Defender.RPG.StrengthCheck(Attacker.Strength, Defender.RPG.StrengthModifiers.ToArray());
+                canBlock = Defender.RPG.EnduranceCheck(Attacker.Strength, Defender.RPG.EnduranceModifiers.ToArray());
             }
 
             // if the defender can't dodge the attack
@@ -65,7 +67,7 @@ namespace LinkEngine.RPG
             {
                 // Deal damage to defender
                 // Defenders endurance and agility has a chance counteract the damage
-                damageAmount = (damage(Attacker.Strength, null) - PlayerBlock(Defender.Endurance, Defender.Agility, Defender.EnduranceModifiers, Defender.AgilityModifiers));
+                damageAmount = (damage(Attacker.Strength, null) - PlayerBlock(Defender.RPG.Endurance, Defender.RPG.Agility, Defender.RPG.EnduranceModifiers, Defender.RPG.AgilityModifiers));
 
                 // Make sure damage amount is not negative, that will give the player health
                 if (damageAmount > 0)
@@ -81,19 +83,19 @@ namespace LinkEngine.RPG
         /// </summary>
         /// <param name="Defender"></param>
         /// <param name="Attacker"></param>
-        public void PlayerAttack (Monster Defender, RPGCharacter Attacker)
+        public void PlayerAttack (Enemy Defender, Player Attacker)
         {
             // Check if the defender has an agility or luck Ability
-            if (Attacker.Luck > 0 || Attacker.Agility > 0)
+            if (Attacker.RPG.Luck > 0 || Attacker.RPG.Agility > 0)
             {
                 // check for any special abilities
                 // if the luck or agility Ability is high enough, Defender can't dodge the attack
-                canDodge = !Attacker.LuckCheck(Defender.Strength, Attacker.LuckModifiers.ToArray());
-                canDodge = !Attacker.AgilityCheck(Defender.Strength, Attacker.AgilityModifiers.ToArray());
+                canDodge = !Attacker.RPG.LuckCheck(Defender.Strength, Attacker.RPG.LuckModifiers.ToArray());
+                canDodge = !Attacker.RPG.AgilityCheck(Defender.Strength, Attacker.RPG.AgilityModifiers.ToArray());
 
                 // if the defender has a higher strength or endurance, he can block the attack
-                canBlock = !Attacker.StrengthCheck(Defender.Strength, Attacker.StrengthModifiers.ToArray());
-                canBlock = !Attacker.EnduranceCheck(Defender.Strength, Attacker.EnduranceModifiers.ToArray());
+                canBlock = !Attacker.RPG.StrengthCheck(Defender.Strength, Attacker.RPG.StrengthModifiers.ToArray());
+                canBlock = !Attacker.RPG.EnduranceCheck(Defender.Strength, Attacker.RPG.EnduranceModifiers.ToArray());
             }
 
             // if the defender can't dodge the attack
@@ -101,7 +103,7 @@ namespace LinkEngine.RPG
             {
                 // Deal damage to defender
                 // Defenders endurance and agility has a chance counteract the damage
-                damageAmount = (damage(Attacker.Strength, null) - MonsterBlock(Defender.Defense));
+                damageAmount = (damage(Attacker.RPG.Strength, null) - MonsterBlock(Defender.Defense));
 
                 // Make sure damage amount is not negative, that will give the player health
                 if (damageAmount > 0)
