+++ LinkEngine/Gameplay/Items/DungeonItem.cs
@@ -1,23 +1,16 @@
 ﻿using LinkEngine.Entities;
+using LinkEngine.Gameplay.Items;
 
 namespace LinkEngine.Dungeon
 {
-    public class DungeonItem : Item
+    public class DungeonItem
     {
         /// <summary>
         /// Bool flag indicating whether this item has been been discovered or is still unknown
         /// </summary>
         public bool HasBeenDiscovered { get; set; }
-
-        /// <summary>
-        /// The Name of the Item. Defaults as "???" until it has been discovered
-        /// </summary>
-        public string ItemName { get; set; }
-        /// <summary>
-        /// The plural Name of the Item. Defaults as "???" until it has been discovered
-        /// </summary>
-        public string ItemNamePlural { get; set; }
-
+        
+        public Item Details { get; set; }
         /// <summary>
         /// The location of the image. image will be changed from game to game to create a randomized mechanic
         /// </summary>
@@ -31,15 +24,9 @@ namespace LinkEngine.Dungeon
         /// <param name="_namePlural"></param>
         /// <param name="equipable"></param>
         /// <param name="consumable"></param>
-        public DungeonItem(int _id, string _name, string _namePlural, bool equipable, bool consumable) : base(_id, _name, _namePlural)
+        public DungeonItem(int _id, string _name, string _namePlural, bool equipable, bool consumable)
         {
             HasBeenDiscovered = false;
-
-            ItemName = "???";
-            ItemNamePlural = "???";
-
-            Equipable = equipable;
-            Consumable = consumable;
         }
     }
 }
