+++ LinkEngine/Gameplay/Items/InventoryItem.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace LinkEngine.Gameplay.Items
+{
+    public class InventoryItem
+    {
+        public Item Details { get; set; }
+        public int Quantity { get; set; }
+        public InventoryItem(Item details, int quantity)
+        {
+            Details = details;
+            Quantity = quantity;
+        }
+    }
+}
