+++ LinkEngine/Gameplay/Items/LootItem.cs
@@ -1,0 +1,16 @@
+﻿namespace LinkEngine.Gameplay.Items
+{
+    public class LootItem
+    {
+        public Item Details { get; set; }
+        public int DropPercentage { get; set; }
+        public bool IsDefaultItem { get; set; }
+
+        public LootItem(Item details, int dropPercentage, bool isDefaultItem)
+        {
+            Details = details;
+            DropPercentage = dropPercentage;
+            IsDefaultItem = isDefaultItem;
+        }
+    }
+}
