+++ LinkEngine/Gameplay/Resource Management/Turn-Based/Faction.cs
