+++ LinkEngine/Gameplay/Resource Management/Turn-Based/Leader.cs
