+++ LinkEngine/Gameplay/Resource Management/Turn-Based/Player.cs
