+++ LinkEngine/Gameplay/Resource Management/Turn-Based/World.cs
