+++ LinkEngine/Gameplay/Survival/Food.cs
@@ -1,9 +1,11 @@
-﻿namespace LinkEngine.Survival
+﻿using LinkEngine.Gameplay.Items;
+
+namespace LinkEngine.Survival
 {
     public class Food
     {
         public short AmountToFeed { get; set; }
-        public Entities.Item Details { get; set; }
+        public Item Details { get; set; }
         public short Quantity { get; set; }
     }
 }
