+++ LinkEngine/LinkEngine.csproj
@@ -20,6 +20,7 @@
     <DefineConstants>DEBUG;TRACE</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
+    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
     <DebugType>pdbonly</DebugType>
@@ -38,65 +39,48 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Components\Animation\Animator2D.cs" />
-    <Compile Include="Components\Animation\Screen.cs" />
+    <Compile Include="Gameplay\Items\CraftingItem.cs" />
+    <Compile Include="Gameplay\Items\InventoryItem.cs" />
+    <Compile Include="Gameplay\Items\LootItem.cs" />
+    <Compile Include="Rendering\2dGraphics\Screen.cs" />
     <Compile Include="Components\Physics\CircleCollider2D.cs" />
     <Compile Include="Components\Physics\Collider2D.cs" />
     <Compile Include="Components\Physics\Transform.cs" />
     <Compile Include="Components\Physics\Vector.cs" />
     <Compile Include="Components\UI\Camera.cs" />
     <Compile Include="Components\Physics\BoxCollider2D.cs" />
-    <Compile Include="Components\UI\Canvas.cs" />
     <Compile Include="Entities\Enemy.cs" />
     <Compile Include="Entities\Entity.cs" />
     <Compile Include="Entities\Player.cs" />
-    <Compile Include="GameTypes\Top-Down\Dungeon\Entities\DungeonEnemy.cs" />
-    <Compile Include="GameTypes\Top-Down\Dungeon\Gameplay\DungeonEquipment.cs" />
-    <Compile Include="GameTypes\Top-Down\Dungeon\Gameplay\DungeonItem.cs" />
-    <Compile Include="GameTypes\Top-Down\Dungeon\Entities\DungeonPlayer.cs" />
-    <Compile Include="GameTypes\Top-Down\Dungeon\Gameplay\DungeonPotion.cs" />
-    <Compile Include="GameTypes\Top-Down\RPG\Engine\Entities\Item.cs" />
-    <Compile Include="GameTypes\Top-Down\Strategy\Turn-Based\Faction.cs" />
-    <Compile Include="GameTypes\Top-Down\Strategy\Turn-Based\Leader.cs" />
-    <Compile Include="GameTypes\Top-Down\Strategy\Turn-Based\Player.cs" />
-    <Compile Include="GameTypes\Top-Down\Strategy\Turn-Based\Territory.cs" />
-    <Compile Include="GameTypes\Top-Down\Strategy\Turn-Based\World.cs" />
+    <Compile Include="Gameplay\Items\DungeonItem.cs" />
+    <Compile Include="Gameplay\Resource Management\Turn-Based\Faction.cs" />
+    <Compile Include="Gameplay\Resource Management\Turn-Based\Leader.cs" />
+    <Compile Include="Gameplay\Resource Management\Turn-Based\Player.cs" />
+    <Compile Include="Gameplay\Resource Management\Turn-Based\Territory.cs" />
+    <Compile Include="Gameplay\Resource Management\Turn-Based\World.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Rendering\2dGraphics\ScreenObject.cs" />
-    <Compile Include="Rendering\2dGraphics\ScreenText.cs" />
     <Compile Include="Components\UI\HUDObject.cs" />
-    <Compile Include="GameTypes\Top-Down\RPG\Engine\Entities\Character.cs" />
-    <Compile Include="GameTypes\Top-Down\RPG\Engine\Entities\Monster.cs" />
-    <Compile Include="GameTypes\Top-Down\RPG\Engine\Entities\NPC.cs" />
-    <Compile Include="GameTypes\Top-Down\RPG\Engine\Entities\PartyMember.cs" />
-    <Compile Include="GameTypes\Top-Down\RPG\Engine\Gameplay\Combat.cs" />
-    <Compile Include="GameTypes\Top-Down\RPG\Engine\Gameplay\Consumable.cs" />
-    <Compile Include="GameTypes\Top-Down\RPG\Engine\Gameplay\Equipment.cs" />
-    <Compile Include="Entities\Item.cs" />
-    <Compile Include="GameTypes\Top-Down\RPG\Engine\Gameplay\Modifier.cs" />
-    <Compile Include="GameTypes\Top-Down\RPG\Engine\Gameplay\Quest.cs" />
-    <Compile Include="GameTypes\Top-Down\RPG\Engine\Gameplay\Ability.cs" />
-    <Compile Include="GameTypes\Top-Down\Survival\Entities\Character.cs" />
-    <Compile Include="GameTypes\Top-Down\Survival\Gameplay\Food.cs" />
+    <Compile Include="Entities\Adventurer.cs" />
+    <Compile Include="Entities\NPC.cs" />
+    <Compile Include="Gameplay\Combat\Combat.cs" />
+    <Compile Include="Gameplay\Items\Consumable.cs" />
+    <Compile Include="Gameplay\Items\Equipment.cs" />
+    <Compile Include="Gameplay\Items\Item.cs" />
+    <Compile Include="Gameplay\Modifiers\Modifier.cs" />
+    <Compile Include="Gameplay\Quests\Quest.cs" />
+    <Compile Include="Gameplay\Skills\Ability.cs" />
+    <Compile Include="Entities\Survivalist.cs" />
+    <Compile Include="Gameplay\Survival\Food.cs" />
     <Compile Include="WorldGen\Modules\Biome.cs" />
-    <Compile Include="GameTypes\Top-Down\Dungeon\Gameplay\Dungeon.cs" />
-    <Compile Include="GameTypes\Top-Down\RPG\Engine\World\Location.cs" />
+    <Compile Include="WorldGen\Modules\Location.cs" />
     <Compile Include="WorldGen\Modules\Chunk.cs" />
     <Compile Include="WorldGen\Generation\DungeonGeneration.cs" />
     <Compile Include="WorldGen\Generation\ProceduralGeneration.cs" />
-    <Compile Include="WorldGen\Modules\Room.cs" />
     <Compile Include="WorldGen\Generation\SetGeneration.cs" />
     <Compile Include="WorldGen\Modules\Tile.cs" />
-    <Compile Include="GameTypes\Top-Down\RPG\Engine\World\World.cs" />
     <Compile Include="WorldGen\Modules\Transition.cs" />
     <Compile Include="WorldGen\World.cs" />
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="GameTypes\Platformer\Beat %27Em Up\" />
-    <Folder Include="GameTypes\Platformer\Fighting\" />
-    <Folder Include="GameTypes\Platformer\Shooter\Shoot %27Em Up\" />
-    <Folder Include="GameTypes\Platformer\Sidescroller\" />
-    <Folder Include="GameTypes\Top-Down\Adventure\" />
-    <Folder Include="GameTypes\Top-Down\Strategy\Real-TIme\" />
-  </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
